Past Order Header Details + Order Inner Details + API Implemented

+++ app/src/main/java/com/atozcorporation/atoz/ui/outlet/outletdetails/OutletDetailsActivity.kt
@@ -1,8 +1,10 @@
 package com.atozcorporation.atoz.ui.outlet.outletdetails
 
+import android.app.AlertDialog
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.widget.Toast
 import com.atozcorporation.atoz.R
 import com.atozcorporation.atoz.base.BaseActivity
 import com.atozcorporation.atoz.base.offlinedb.Order_Summery_Db_Helper
@@ -10,6 +12,7 @@ import com.atozcorporation.atoz.rest.response.login.LoginResponse
 import com.atozcorporation.atoz.rest.response.outlet.OutletListResponse
 import com.atozcorporation.atoz.ui.addoutlet.AddOutletActivity
 import com.atozcorporation.atoz.ui.manageproduct.productcategorylist.ProductCategoryActivity
+import com.growinginfotech.businesshub.base.defaultToast
 import com.growinginfotech.businesshub.base.navigateTo
 import kotlinx.android.synthetic.main.activity_outlet_details.*
 
@@ -39,16 +42,36 @@ class OutletDetailsActivity : BaseActivity() {
             startActivity(intent)
         }
         buttonPlaceOrder.setOnClickListener {
-           val dbHelper = Order_Summery_Db_Helper(this, null);
-            dbHelper.deleteAll()
-            val orderForUser = LoginResponse.UserDetails()
-            orderForUser.id = outletDetails.id
-            orderForUser.batchId = outletDetails.batchId
-            orderForUser.name = outletDetails.name
-            orderForUser.personName = outletDetails.personName
-            orderForUser.contactNumber = outletDetails.contactNumber
-            setOrderForUser(orderForUser)
-            navigateTo<ProductCategoryActivity> {  }
+            if(orderFor?.id != outletDetails.id){
+                val builder = AlertDialog.Builder(this)
+                builder.setTitle("Title")
+                builder.setMessage("Details")
+                builder.setIcon(android.R.drawable.ic_dialog_alert)
+
+                //performing positive action
+                builder.setPositiveButton("Yes"){dialogInterface, which ->
+                    Toast.makeText(applicationContext,"clicked yes",Toast.LENGTH_LONG).show()
+                    val dbHelper = Order_Summery_Db_Helper(this, null);
+                    dbHelper.deleteAll()
+                    val orderForUser = LoginResponse.UserDetails()
+                    orderForUser.id = outletDetails.id
+                    orderForUser.batchId = outletDetails.batchId
+                    orderForUser.name = outletDetails.name
+                    orderForUser.personName = outletDetails.personName
+                    orderForUser.contactNumber = outletDetails.contactNumber
+                    setOrderForUser(orderForUser)
+                    navigateTo<ProductCategoryActivity> {  }
+                }
+                //performing negative action
+                builder.setNegativeButton("No"){dialogInterface, which ->
+                    //
+                }
+                val alertDialog: AlertDialog = builder.create()
+                alertDialog.setCancelable(false)
+                alertDialog.show()
+            } else {
+                navigateTo<ProductCategoryActivity> {  }
+            }
         }
 
         textViewBusinessName.text = outletDetails.name
